pyproject.toml
[tool.poetry]
name = "aztest1"
version = "0.1.0"
description = ""
authors = ["Fran Hrzenjak <fran.hrzenjak@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
dash = "^1.18.1"
pandas-datareader = "^0.9.0"
dash-bootstrap-components = "^0.11.1"
casadi = "^3.5.5"
AeroSandbox = "^2.3.5"
pytest = "^6.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
